Fix bug where __ccContentScaleFactor was re-declared as double due to CGFloat typedef in Xcode 5.1.  Add copy of podsepc with source changed to point to fork

+++ cocos2d/Platforms/iOS/CCDirectorIOS.h
@@ -77,7 +77,7 @@
 {
 	/* contentScaleFactor could be simulated */
 	BOOL	_isContentScaleSupported;
-	
+
 	CCTouchDispatcher	*_touchDispatcher;
 }
 
@@ -106,6 +106,6 @@
 @end
 
 // optimization. Should only be used to read it. Never to write it.
-extern CGFloat	__ccContentScaleFactor;
+extern float	__ccContentScaleFactor;
 
 #endif // __CC_PLATFORM_IOS
